create shared GetMockClient - wip
