+++ src/ui/connect-ui/tests/FamilyHubs.ReferralUi.UnitTests/Services/BaseClientService.cs
@@ -8,21 +8,6 @@ namespace FamilyHubs.ReferralUi.UnitTests.Services;
 
 public class BaseClientService
 {
-    protected static HttpClient GetMockClient(string content)
-    {
-        var mockHttpMessageHandler = new Mock<HttpMessageHandler>();
-        mockHttpMessageHandler.Protected()
-            .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
-            .ReturnsAsync(new HttpResponseMessage
-            {
-                Content = new StringContent(content),
-                StatusCode = HttpStatusCode.OK
-            });
-
-        var client = new HttpClient(mockHttpMessageHandler.Object);
-        client.BaseAddress = new Uri("Https://Localhost");
-        return client;
-    }
 
     protected OrganisationDetailsDto GetTestCountyCouncilDto()
     {
